+++ core/authority_discovery/CMakeLists.txt
@@ -3,5 +3,5 @@
 # SPDX-License-Identifier: Apache-2.0
 #
 
-add_subdirectory(impl)
+add_subdirectory(publisher)
 add_subdirectory(protobuf)
