+++ core/authority_discovery/publisher/CMakeLists.txt
@@ -4,7 +4,7 @@
 #
 
 add_library(address_publisher
-        address_publisher_impl.cpp
+        address_publisher.cpp
         )
 target_link_libraries(address_publisher
         logger
